+++ ImperitWASM/Client/Services/SessionStorage.cs
@@ -8,12 +8,12 @@ namespace ImperitWASM.Client.Services
 		readonly ISessionStorageService iss;
 		public Data.Session Session { get; private set; } = new Data.Session();
 		public int? PlayerOrder { get; private set; } = null;
-		public long? GameId { get; private set; } = null;
+		public int? GameId { get; private set; } = null;
 		public SessionStorage(ISessionStorageService iss) => this.iss = iss;
 		public async Task LoadAsync() => Session = await iss.GetItemAsync<Data.Session?>("session") ?? new Data.Session();
 		public Task LoginAsync(Data.LoginResult login)
 		{
-			(PlayerOrder, GameId) = login.S.IsSet() ? (login.I, login.G) : (null as int?, null as long?);
+			(PlayerOrder, GameId) = login.S.IsSet() ? (login.I, login.G) : (null as int?, null as int?);
 			return ResetAsync(login.S);
 		}
 		public Task ResetAsync(Data.Session? session = null)
