+++ ImperitWASM/Server/Services/ProvinceLoader.cs
@@ -11,8 +11,8 @@ namespace ImperitWASM.Server.Services
 {
 	public interface IProvinces
 	{
-		Provinces this[long gameId] { get; }
-		Province this[long gameId, int regionId] { get; }
+		Provinces this[int gameId] { get; }
+		Province this[int gameId, int regionId] { get; }
 		Task AddAsync(IEnumerable<Province> provinces);
 		Task UpdateAsync(IEnumerable<Province> provinces);
 	}
@@ -25,8 +25,8 @@ namespace ImperitWASM.Server.Services
 			this.ctx = ctx;
 			this.graph = graph;
 		}
-		public Provinces this[long gameId] => new Provinces(ctx.Provinces!.AsNoTracking().Where(province => province.GameId == gameId).OrderBy(province => province.RegionId).AsEnumerable().ToImmutableArray(), graph);
-		public Province this[long gameId, int regionId] => ctx.Provinces!.AsNoTracking().Single(province => province.GameId == gameId && province.RegionId == regionId);
+		public Provinces this[int gameId] => new Provinces(ctx.Provinces!.AsNoTracking().Where(province => province.GameId == gameId).OrderBy(province => province.RegionId).AsEnumerable().ToImmutableArray(), graph);
+		public Province this[int gameId, int regionId] => ctx.Provinces!.AsNoTracking().Single(province => province.GameId == gameId && province.RegionId == regionId);
 
 		public Task AddAsync(IEnumerable<Province> provinces)
 		{
